# swagger/template.yml
swagger: "2.0"
info:
  title: edu_forms API
  description: Multi-tenant educational assessment platform
  version: "1.0.0"
host: "127.0.0.1:5000"
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "JWT access token. Example: Bearer eyJhbG..."
  WorkspaceId:
    type: apiKey
    name: X-Workspace-Id
    in: header
    description: Active workspace context (not in JWT)
...
paths:
...
  /workspaces:
    get:
      tags: [Workspaces]
      summary: List accessible workspaces
      security: [{ Bearer: [] }]
      responses:
        200:
          description: Workspace list
    post:
      tags: [Workspaces]
      summary: Create workspace
      security: [{ Bearer: [] }]
      parameters:
        - in: body
          name: body
          schema:
            type: object
            properties:
              name: { type: string }
              kind: { type: string, enum: [SOLO, INSTITUTION] }
              slug: { type: string }
              logo_url: { type: string, example: "https://cdn.example.com/logos/school.png" }
              description: { type: string, example: "Independent teacher workspace" }
      responses:
        201:
          description: Created
  /workspaces/teachers:
    get:
      tags: [Workspaces]
      summary: List institution workspace teachers
      description: |
        Returns active members with role TEACHER in the workspace selected via X-Workspace-Id.
        Institution workspaces only. Caller must be workspace owner or ADMIN.
        Teachers and students receive 403.
        Supports optional `search` (full_name, email), `page`, and `per_page`.
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: search
          in: query
          type: string
          required: false
        - name: page
          in: query
          type: integer
          default: 1
        - name: per_page
          in: query
          type: integer
          default: 20
      responses:
        200:
          description: Teacher list
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              count: { type: integer, description: "Items in current page" }
              total: { type: integer }
              page: { type: integer }
              per_page: { type: integer }
              pages: { type: integer }
              teachers:
                type: array
                items:
                  $ref: "#/definitions/WorkspaceTeacherMember"
              data:
                type: array
                description: Backward-compatible alias of teachers
                items:
                  $ref: "#/definitions/WorkspaceTeacherMember"
        403:
          description: Forbidden (not owner/admin, or not an institution workspace)
    delete:
      tags: [Workspaces]
      summary: Remove teacher from institution workspace
      description: |
        Removes a TEACHER membership from the workspace selected via X-Workspace-Id.
        Institution workspaces only. Caller must be workspace owner or ADMIN.

        **Swagger:** run `GET /workspaces/teachers` first, copy `membership_id` from a teacher
        row, then pass it as the query parameter below.

        Cleans up teacher subject assignments in this workspace, then deletes the membership.
        Does not delete the user account or memberships in other workspaces.
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: membership_id
          in: query
          type: integer
          required: true
          example: 45
          description: Workspace membership id of the teacher (from GET /workspaces/teachers)
      responses:
        200:
          description: Teacher removed
          schema:
            type: object
            properties:
              message:
                type: string
                example: Teacher removed from workspace successfully.
        403:
          description: Forbidden (not owner/admin, cannot remove owner)
        404:
          description: Membership not found in this workspace
        422:
          description: Membership is not a teacher or not active
  /workspaces/students:
    get:
      tags: [Workspaces]
      summary: List workspace students
      description: |
        Returns students in the workspace selected via X-Workspace-Id.

        **INSTITUTION:** all active STUDENT memberships, with `enrolled_subjects_count`
        within the institution.

        **SOLO (independent teacher):** students enrolled in at least one subject
        in this workspace, with `enrolled_subjects_count` for subjects in this workspace.

        Caller must be workspace owner or ADMIN. Teachers endpoint remains institution-only.
        Supports optional `search` (full_name, email), `page`, and `per_page`.
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: search
          in: query
          type: string
          required: false
        - name: page
          in: query
          type: integer
          default: 1
        - name: per_page
          in: query
          type: integer
          default: 20
      responses:
        200:
          description: Student list
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              count: { type: integer }
              total: { type: integer }
              page: { type: integer }
              per_page: { type: integer }
              pages: { type: integer }
              students:
                type: array
                items:
                  $ref: "#/definitions/WorkspaceStudentMember"
        403:
          description: Forbidden (not owner/admin, or not an institution workspace)
    delete:
      tags: [Workspaces]
      summary: Remove student from workspace
      description: |
        Removes a STUDENT membership from the workspace selected via X-Workspace-Id.
        Institution or SOLO workspace. Caller must be workspace owner or ADMIN.

        **Swagger:** run `GET /workspaces/students` first, copy `membership_id` from a student
        row, then pass it as the query parameter below.

        Cleans up subject enrollments, student group memberships, and test assignments
        in this workspace, then deletes the membership.
        Does not delete the user account or memberships in other workspaces.
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: membership_id
          in: query
          type: integer
          required: true
          example: 45
          description: Workspace membership id of the student (from GET /workspaces/students)
      responses:
        200:
          description: Student removed
          schema:
            type: object
            properties:
              message:
                type: string
                example: Student removed from workspace successfully.
        403:
          description: Forbidden (not owner/admin, cannot remove owner)
        404:
          description: Membership not found in this workspace
        422:
          description: Membership is not a student or not active
  /workspaces/members/recently-active:
    get:
      tags: [Workspaces]
      summary: List recently active workspace members
      description: |
        Returns students and/or teachers in the workspace selected via `X-Workspace-Id`
        who were active within the last 24 hours.

        Supports **INSTITUTION** and **SOLO** workspaces. SOLO is an independent teacher
        workspace; the owner is treated as the workspace administrator.

        Activity is derived from `TestAttempt.last_activity_at` in this workspace, falling back to
        `User.last_login_at`. Caller must be workspace owner or ADMIN.

        Optional `role` query parameter filters to `STUDENT` or `TEACHER`. When omitted, both
        roles are returned (ADMIN memberships are excluded).
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: role
          in: query
          type: string
          required: false
          enum: [STUDENT, TEACHER]
          description: Filter by workspace membership role
      responses:
        200:
          description: Recently active members
          schema:
            type: object
            properties:
              members:
                type: array
                items:
                  type: object
                  properties:
                    membership_id: { type: integer, example: 45 }
                    user_id: { type: integer, example: 14 }
                    full_name: { type: string, example: "Ahmed Ali" }
                    email: { type: string, example: "ahmed@example.com" }
                    avatar: { type: string, example: "https://cdn.example.com/avatars/user123.png" }
                    role: { type: string, enum: [STUDENT, TEACHER], example: STUDENT }
                    status: { type: string, example: ACTIVE }
                    last_activity_at: { type: string, format: date-time }
        403:
          description: Forbidden (not owner/admin, or unsupported workspace type)
  /workspaces/members/{membership_id}:
    get:
      tags: [Workspaces]
      summary: Get workspace member details
      description: |
        Returns detail for a **STUDENT** or **TEACHER** member in the workspace
        selected via `X-Workspace-Id`. Role is determined from the membership record.

        Supports **INSTITUTION** and **SOLO** workspaces. SOLO is an independent teacher
        workspace; the owner is treated as the workspace administrator.

        Caller must be workspace owner or ADMIN. Membership must belong to the
        current workspace.

        **Student response:** `student`, `subjects`, `upcoming_tests`, `completed_tests`.

        **Teacher response:** `teacher`, `subjects`, `statistics` (question banks and tests created).
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: membership_id
          in: path
          type: integer
          required: true
          example: 45
          description: Workspace membership id (from GET /workspaces/teachers or /workspaces/students)
      responses:
        200:
          description: Member details (student or teacher shape)
          schema:
            type: object
            properties:
              student:
                type: object
                description: Present when membership role is STUDENT
                properties:
                  user_id: { type: integer }
                  full_name: { type: string }
                  email: { type: string }
                  avatar: { type: string }
                  last_activity_at: { type: string, format: date-time }
                  status: { type: string, example: ACTIVE }
              teacher:
                type: object
                description: Present when membership role is TEACHER
                properties:
                  user_id: { type: integer }
                  full_name: { type: string }
                  email: { type: string }
                  avatar: { type: string }
                  last_activity_at: { type: string, format: date-time }
                  status: { type: string, example: ACTIVE }
              subjects:
                type: array
                items:
                  type: object
                  properties:
                    subject_id: { type: integer }
                    subject_name: { type: string }
              upcoming_tests:
                type: array
                description: Student only
                items:
                  type: object
                  properties:
                    test_id: { type: integer }
                    test_name: { type: string }
                    subject_name: { type: string }
                    availability_type: { type: string, enum: [FLEXIBLE, SCHEDULED] }
                    starts_at: { type: string, format: date-time }
                    duration_minutes: { type: integer }
              completed_tests:
                type: array
                description: Student only
                items:
                  type: object
                  properties:
                    test_id: { type: integer }
                    test_name: { type: string }
                    test_type: { type: string, enum: [FLEXIBLE, SCHEDULED] }
                    test_date: { type: string, format: date }
                    test_start_time: { type: string, format: date-time }
                    subject_name: { type: string }
                    duration_minutes: { type: integer }
              statistics:
                type: object
                description: Teacher only
                properties:
                  question_banks_created: { type: integer }
                  tests_created: { type: integer }
        403:
          description: Forbidden (not owner/admin, or unsupported workspace type)
        404:
          description: Membership not found in this workspace
        422:
          description: Membership is not a student or teacher
    patch:
      tags: [Workspaces]
      summary: Update workspace member profile
      description: |
        Updates editable **User** profile fields for a workspace member identified by `membership_id`.
        Requires `X-Workspace-Id`. Caller must be workspace owner or ADMIN.

        Applies to any active member role (TEACHER, STUDENT, ADMIN) in the current workspace.
        Only updates linked User fields; does **not** change `membership_id`, `workspace_id`, `role`, or `user_id`.

        **Swagger:** run `GET /workspaces/teachers` or `GET /workspaces/students`, copy `membership_id`,
        then call this endpoint with that id in the path.

        Editable fields: `full_name`, `phone_number`, `avatar_url` (stored as `profile_image_url`).
        Email is read-only and cannot be changed via this endpoint.
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: membership_id
          in: path
          type: integer
          required: true
          example: 45
          description: Workspace membership id (from teachers/students list)
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              full_name: { type: string, example: "Ahmed Ali" }
              phone_number: { type: string, example: "+9639xxxxxxx" }
              avatar_url: { type: string, example: "https://cdn.example.com/avatars/user123.png" }
      responses:
        200:
          description: Member updated
          schema:
            type: object
            properties:
              message:
                type: string
                example: Workspace member updated successfully
              member:
                type: object
                description: Same shape as teacher/student list items
                properties:
                  user_id: { type: integer, example: 14 }
                  membership_id: { type: integer, example: 45 }
                  full_name: { type: string, example: "Ahmed Ali" }
                  email: { type: string, example: "ahmed@example.com" }
                  avatar_url: { type: string }
                  user_status: { type: string, example: ACTIVE }
                  workspace_role: { type: string, example: TEACHER }
                  membership_role: { type: string, example: TEACHER }
                  assigned_subjects_count: { type: integer, example: 4 }
                  enrolled_subjects_count: { type: integer, example: 5 }
                  created_at: { type: string, format: date-time }
        403:
          description: Forbidden (not owner/admin)
        404:
          description: Membership or user not found in this workspace
        422:
          description: Validation error (empty body, inactive membership, invalid field)
  /workspaces/{workspace_id}:
    get:
      tags: [Workspaces]
      summary: Get workspace
      security: [{ Bearer: [] }]
      parameters:
        - name: workspace_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: Workspace detail
    patch:
      tags: [Workspaces]
      summary: Update workspace
      security: [{ Bearer: [] }]
      parameters:
        - name: workspace_id
          in: path
          type: integer
          required: true
        - in: body
          name: body
          schema:
            type: object
            properties:
              name: { type: string }
              slug: { type: string }
              logo_url: { type: string, example: "https://cdn.example.com/logos/school.png" }
              description: { type: string, example: "Independent teacher workspace" }
      responses:
        200:
          description: Updated
    delete:
      tags: [Workspaces]
      summary: Delete workspace (owner only)
      security: [{ Bearer: [] }]
      parameters:
        - name: workspace_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: Deleted
...
  /question-banks/workspace:
    get:
      tags: [Question banks]
      summary: Discover workspace question banks (other teachers)
      description: |
        Returns active WORKSPACE visibility banks created by other members in the
        current workspace. Excludes banks created by the current user.
        Teachers see banks only for subjects they teach; workspace owner/admin see
        banks across all workspace subjects.
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: page
          in: query
          type: integer
          default: 1
        - name: per_page
          in: query
          type: integer
          default: 20
      responses:
        200:
          description: Paginated workspace banks
          schema:
            type: object
            properties:
              question_banks: { type: array, items: { type: object } }
              count: { type: integer }
              total: { type: integer }
              page: { type: integer }
              per_page: { type: integer }
              pages: { type: integer }
...
definitions:
...
  WorkspaceTeacherMember:
    type: object
    properties:
      user_id: { type: integer, example: 14 }
      membership_id: { type: integer, example: 51 }
      full_name: { type: string, example: "Ahmed Ali" }
      email: { type: string, example: "ahmed@example.com" }
      avatar_url: { type: string }
      user_status: { type: string, example: "ACTIVE" }
      workspace_role: { type: string, example: "TEACHER" }
      membership_role: { type: string, example: "TEACHER" }
      assigned_subjects_count: { type: integer, example: 4 }
      created_at: { type: string, format: date-time }
  WorkspaceStudentMember:
    type: object
    properties:
      user_id: { type: integer, example: 21 }
      membership_id: { type: integer, example: 84 }
      full_name: { type: string, example: "Sara Mohamed" }
      email: { type: string, example: "sara@example.com" }
      avatar_url: { type: string }
      user_status: { type: string, example: "ACTIVE" }
      workspace_role: { type: string, example: "STUDENT" }
      enrolled_subjects_count: { type: integer, example: 5 }
...